+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## Improvements
 
+- Add support for WebSockets in RPC.
+
 # 0.5.0
 
 ## Bug Fixes
@@ -15,6 +17,10 @@
 - Forbidden CALLCODE and DELEGATECALL to internal contracts.
 
 - RPC now returns the correct rlp size of the block
+
+- Fix a race condition that may cause optimistic execution to panic.
+
+- Delay block requests when we cannot process them to avoid wasting network bandwidth.
 
 ## Improvements
 
@@ -41,6 +47,10 @@
 - Fix definitions and logics in transaction early execution error checking.
 
 - Use block_count - 1 in target difficulty calculation because it's the unbiased estimation of exponential distribution parameter (past mining power).
+
+- Add fields in Receipt: gas_fee, gas_sponsored, storage_sponsored. Accumulate gas_used in Receipt, not gas_charged.
+
+- Define gas_used to be transaction gas limit for NotEnoughCash, the same as all other exceptions.
 
 ## Improvements